Shrink the bbox before constructing a DCR-matched template.
The DCR model creates artifacts over a few pixels at the edges of each patch. Patches have sufficient overlap to negate this, but when the `coaddPatch` is added to the `coaddExposure` the entire patch out to the outer bounding box is inserted, over-writing any data already there. This means that the last patch added will include artifacts at the edge. One possible fix is to shrink the outer bounding boxes by a few pixels, so that the edge of the overlap region is cut. This prevents the later patches from overwriting good pixels in the boundary region with data contaminated by artifacts.